moved styled components into separate file
