+++ frontend/src/pages/SignIn.js
@@ -1,106 +1,25 @@
 import React, { useState, useEffect } from "react";
 import { useSelector, useDispatch, batch } from "react-redux";
 import { useHistory } from "react-router-dom";
-import styled from "styled-components";
 
 import user from "../reducers/user";
-
 import { API_URL } from "../reusables/urls";
-
 import { Loading } from "../components/Loading";
-import { Button } from "components/Button";
+import { Button } from "components/Buttons";
 import { InputForm } from "components/InputForm";
 import { Footer } from "components/Footer";
 import logo from "assets/logonew.png";
 import pic from "assets/pic.jpg";
+import { 
+  SignInUpContainer, 
+  FormContainer, 
+  SignInLogo, 
+  SignInUpForm, 
+  ErrorMessage, 
+  EyeButton, 
+  FormImageContainer, 
+  FormImage } from '../components/StylingPages';
 
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  align-items: center;
-  height: 50%;
-  width: 100%;
-  margin-bottom: 100px;
-  @media (min-width: 768px) {
-    flex-direction: row;
-    height: 60%;
-    max-width: 75%;
-    box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
-  }
-  @media (min-width: 1200px) {
-    width: 60%;
-    height: 70%;
-  }
-`;
-
-const FormContainer = styled.div`
-  width: 75%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  padding: 30px;
-  align-items: center;
-  @media (min-width: 768px) {
-    height: 100%;
-  }
-`;
-
-const ImageContainer = styled.div`
-  display: none;
-  @media (min-width: 768px) {
-    display: flex;
-    width: 100%;
-    height: 100%;
-  }
-`;
-
-const Image = styled.img`
-  object-fit: cover;
-  width: 100%;
-`;
-
-const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  position: relative;
-`;
-
-const Logo = styled.img`
-  background-image: url(./assets/logo.png);
-  width: 50%;
-  margin-bottom: 100px;
-  @media (min-width: 768px) {
-    margin-bottom: 0;
-  }
-  @media (min-width: 1200px) {
-    width: 40%;
-  }
-`;
-
-const ErrorMessage = styled.p`
-  color: red;
-  margin-top: 10px;
-  font-size: 12px;
-  position: absolute;
-  bottom: 35%;
-`;
-
-const EyeButton = styled.button`
-  background-color: transparent;
-  border: none;
-  outline: none;
-  cursor: pointer;
-  color: #6c6c6d;
-  position: absolute;
-  right: 0;
-  bottom: 59%;
-  :hover {
-    opacity: 0.8;
-  }`
 
 export const SignIn = () => {
   const [username, setUsername] = useState("");
@@ -109,13 +28,11 @@ export const SignIn = () => {
   const [loading, setLoading] = useState(false);
   const [errorMessage, setErrorMessage] = useState("");
   const [showPassword, setShowPassword] = useState(true);
-
   const accessToken = useSelector((store) => store.user.accessToken);
   const dispatch = useDispatch();
   const history = useHistory();
 
   useEffect(() => {
-    // redirect user to '/' path
     console.log("Checking access token", accessToken);
     if (accessToken) {
       history.push("/");
@@ -163,10 +80,10 @@ export const SignIn = () => {
     <>
       {loading && <Loading loadingText="On your way to the jokes.." />}
       {!loading && (
-        <Container>
+        <SignInUpContainer>
           <FormContainer>
-              <Logo src={logo}></Logo>
-            <Form onSubmit={handleFormSubmit}>
+              <SignInLogo src={logo}></SignInLogo>
+            <SignInUpForm onSubmit={handleFormSubmit}>
               <InputForm
                 onChange={(e) => setUsername(e.target.value)}
                 id="username"
@@ -186,12 +103,12 @@ export const SignIn = () => {
               </EyeButton>
               <ErrorMessage>{errorMessage}</ErrorMessage>
               <Button onClick={() => setMode("signin")} buttonText="sign in" />
-            </Form>
+            </SignInUpForm>
           </FormContainer>
-          <ImageContainer>
-            <Image src={pic}></Image>
-          </ImageContainer>
-        </Container>
+          <FormImageContainer>
+            <FormImage src={pic}></FormImage>
+          </FormImageContainer>
+        </SignInUpContainer>
       )}
       <Footer
         footerText="Don't have an account?"
