+++ frontend/src/reusables/urls.js
@@ -1,3 +1,3 @@
-const BASE_URL = "http://localhost:8080";
+const BASE_URL = "https://jokeio.herokuapp.com";
 
 export const API_URL = (slug) => `${BASE_URL}/${slug}`;
